+++ graphDataLib/include/graphOps.h
@@ -77,7 +77,7 @@ typedef struct edge_t * (*funcGetEdges)(const size_t *nodeid, const struct graph
  * @return 0 if there was a problem retrieving the value (such as the edge not existing); otherwise, 1 for a successful
  * retrieval
  */
-typedef int (*funcGetCapacity)(const size_t *uid, const size_t *vid, double *cap, const struct graph_t *graph);
+typedef short (*funcGetCapacity)(const size_t *uid, const size_t *vid, double *cap, const struct graph_t *graph);
 
 /**
  * @brief Function pointer to retrieve the current flow value for a given edge.
@@ -91,7 +91,7 @@ typedef int (*funcGetCapacity)(const size_t *uid, const size_t *vid, double *cap
  * @return 0 if there was a problem retrieving the value (such as the edge not existing); otherwise, 1 for a successful
  * retrieval
  */
-typedef int (*funcGetFlow)(const size_t *uid, const size_t *vid, double *flow, const struct graph_t *graph);
+typedef short (*funcGetFlow)(const size_t *uid, const size_t *vid, double *flow, const struct graph_t *graph);
 
 //Write functions to modify graph
 /**
@@ -101,7 +101,7 @@ typedef int (*funcGetFlow)(const size_t *uid, const size_t *vid, double *flow, c
  * @param g Graph structure to add the node
  * @return 0 if there was an error, 1 if the node was successfully added
  */
-typedef int (*funcAddNode)(const size_t *nodeid, struct graph_t *graph);
+typedef short (*funcAddNode)(const size_t *nodeid, struct graph_t *graph);
 
 /**
  * @brief Remove a node from the graph.
@@ -111,7 +111,7 @@ typedef int (*funcAddNode)(const size_t *nodeid, struct graph_t *graph);
  * @param g Graph structure in question
  * @return 0 if there was an error (node already exists or outside the bounds of the implementation); otherwise, 1 if successful.
  */
-typedef int (*funcRemoveNode)(const size_t *nodeid, struct graph_t *graph);
+typedef short (*funcRemoveNode)(const size_t *nodeid, struct graph_t *graph);
 
 /**
  * @brief Function pointer to add an edge to a given graph.
@@ -122,7 +122,7 @@ typedef int (*funcRemoveNode)(const size_t *nodeid, struct graph_t *graph);
  * @param g graph structure in question
  * @return 0 if there was an error; 1 if the edge was successfully added.
  */
-typedef int (*funcAddEdge)(const size_t *uid, const size_t *vid, double *cap, struct graph_t *graph);
+typedef short (*funcAddEdge)(const size_t *uid, const size_t *vid, double *cap, struct graph_t *graph);
 
 /**
  * @brief Function pointer to remove an edge from the given graph.
@@ -134,7 +134,7 @@ typedef int (*funcAddEdge)(const size_t *uid, const size_t *vid, double *cap, st
  * @param g Graph structure in question
  * @return 0 if there was an error (e.g. the edge was not found); otherwise, 1 if the edge was removed.
  */
-typedef int (*funcRemoveEdge)(const size_t *uid, const size_t *vid, struct graph_t *graph);
+typedef short (*funcRemoveEdge)(const size_t *uid, const size_t *vid, struct graph_t *graph);
 
 /**
  * @brief Function pointer to set the capacity (cost, weight, etc.) of an edge in the given graph.
@@ -144,7 +144,7 @@ typedef int (*funcRemoveEdge)(const size_t *uid, const size_t *vid, struct graph
  * @param g Graph structure in question
  * @return 0 if there was an error; 1 if the capacity was successfully set
  */
-typedef int (*funcSetCapacity)(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
+typedef short (*funcSetCapacity)(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
 
 /**
  * @brief Function to add (adjust) the capacity for an edge by a given amount.
@@ -157,7 +157,7 @@ typedef int (*funcSetCapacity)(const size_t *uid, const size_t *vid, const doubl
  * @param g Graph structure in question
  * @return 0 if there was an error (edge not found, for example); 1 of capacity was successfully adjusted
  */
-typedef int (*funcAddCapacity)(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
+typedef short (*funcAddCapacity)(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
 
 /**
  * @brief Function to set the flow value for an edge (amount of capacity currently "used")
@@ -170,7 +170,7 @@ typedef int (*funcAddCapacity)(const size_t *uid, const size_t *vid, const doubl
  * @param g Graph structure in question
  * @return 0 of there was an error (edge not found, for example); otherwise, 1 if the flow value as successfully set.
  */
-typedef int (*funcSetFlow)(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
+typedef short (*funcSetFlow)(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
 
 /**
  * @brief Function to adjust the flow value of a given edge.
@@ -183,7 +183,7 @@ typedef int (*funcSetFlow)(const size_t *uid, const size_t *vid, const double *f
  * @param g The graph structure in question
  * @return 0 if there was an error (such as the edge not found); otherwise, 1 if the flow value was successfully adjusted.
  */
-typedef int (*funcAddFlow)(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
+typedef short (*funcAddFlow)(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
 
 /**
  * @brief Function pointer to "reset" the graph according to the given argument pointer.
@@ -195,7 +195,7 @@ typedef int (*funcAddFlow)(const size_t *uid, const size_t *vid, const double *f
  * @param callback Callback to be executed when graph has been reset.
  * @return 0 if there was an error during the reset; 1 if the reset completed;
  */
-typedef int (*funcResetGraph)(struct graph_t *graph, void *args, void (*callback)(void));
+typedef short (*funcResetGraph)(struct graph_t *graph, void *args, void (*callback)(void));
 
 
 /**
