+++ graphDataLib/include/impl/array/arraygraph.h
@@ -47,7 +47,7 @@ struct arraydata_t {
  * @param lblcount Number of labels to be used--may be zero, depending on the graph domain.
  * @return 1 if successful; 0 if an error
  */
-int arrayGraphInit(struct graph_t *g);
+short arrayGraphInit(struct graph_t *g);
 
 
 /**
@@ -55,6 +55,6 @@ int arrayGraphInit(struct graph_t *g);
  * @param g graph_t with array structures to be deallocated
  * @return 1 if successful; otherwise, 0.
  */
-int arrayGraphFree(struct graph_t *g);
+short arrayGraphFree(struct graph_t *g);
 
 #endif //GRAPHDATA_ARRAYGRAPH_H
