+++ graphDataLib/include/impl/array/arrayops.h
@@ -39,7 +39,7 @@ struct node_t * arrayGetNode(const size_t *nodeid, const struct graph_t *g);
  * @param g Graph structure in question
  * @return 0 if there was an error (node already exists or outside the bounds of the implementation); otherwise, 1 if successful.
  */
-int arrayRemoveNode(const size_t *nodeid, struct graph_t *g);
+short arrayRemoveNode(const size_t *nodeid, struct graph_t *g);
 /**
  * @brief Implementation to retrieve a edge structure reference.
  *
@@ -81,7 +81,7 @@ struct edge_t * arrayGetEdges(const size_t *nodeid, const struct graph_t *g);
  * @return 0 if there was a problem retrieving the value (such as the edge not existing); otherwise, 1 for a successful
  * retrieval
  */
-int arrayGetCapacity(const size_t *uid, const size_t *vid, double *cap, const struct graph_t *g);
+short arrayGetCapacity(const size_t *uid, const size_t *vid, double *cap, const struct graph_t *g);
 
 /**
  * @brief Function pointer to retrieve the current flow value for a given edge.
@@ -95,7 +95,7 @@ int arrayGetCapacity(const size_t *uid, const size_t *vid, double *cap, const st
  * @return 0 if there was a problem retrieving the value (such as the edge not existing); otherwise, 1 for a successful
  * retrieval
  */
-int arrayGetFlow(const size_t *uid, const size_t *vid, double *flow, const struct graph_t *g);
+short arrayGetFlow(const size_t *uid, const size_t *vid, double *flow, const struct graph_t *g);
 
 
 //Write functions to modify graph
@@ -106,7 +106,7 @@ int arrayGetFlow(const size_t *uid, const size_t *vid, double *flow, const struc
  * @param g Graph structure to add the node
  * @return 0 if there was an error, 1 if the node was successfully added
  */
-int arrayAddNode(const size_t *nodeid, struct graph_t *g);
+short arrayAddNode(const size_t *nodeid, struct graph_t *g);
 /**
  * @brief Implementation to add an edge to a given graph.
  * Not all implementations may use this (for example, fixed-size ARRAY implementations representing a set domain of nodes and relationships).
@@ -116,7 +116,7 @@ int arrayAddNode(const size_t *nodeid, struct graph_t *g);
  * @param graph graph structure in question
  * @return 0 if there was an error; 1 if the edge was successfully added.
  */
-int arrayAddEdge(const size_t *uid, const size_t *vid, double *cap, struct graph_t *graph);
+short arrayAddEdge(const size_t *uid, const size_t *vid, double *cap, struct graph_t *graph);
 
 /**
  * @brief Function pointer to remove an edge from the given graph.
@@ -128,7 +128,7 @@ int arrayAddEdge(const size_t *uid, const size_t *vid, double *cap, struct graph
  * @param graph Graph structure in question
  * @return 0 if there was an error (e.g. the edge was not found); otherwise, 1 if the edge was removed.
  */
-int arrayRemoveEdge(const size_t *uid, const size_t *vid, struct graph_t *graph);
+short arrayRemoveEdge(const size_t *uid, const size_t *vid, struct graph_t *graph);
 
 /**
  * @brief Implementation to set the capacity (cost, weight, etc.) of an edge in the given graph.
@@ -138,7 +138,7 @@ int arrayRemoveEdge(const size_t *uid, const size_t *vid, struct graph_t *graph)
  * @param graph Graph structure in question
  * @return 0 if there was an error; 1 if the capacity was successfully set
  */
-int arraySetCapacity(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
+short arraySetCapacity(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
 
 /**
  * @brief Function to add (adjust) the capacity for an edge by a given amount.
@@ -151,7 +151,7 @@ int arraySetCapacity(const size_t *uid, const size_t *vid, const double *cap, st
  * @param graph Graph structure in question
  * @return 0 if there was an error (edge not found, for example); 1 of capacity was successfully adjusted
  */
-int arrayAddCapacity(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
+short arrayAddCapacity(const size_t *uid, const size_t *vid, const double *cap, struct graph_t *graph);
 
 
 /**
@@ -165,7 +165,7 @@ int arrayAddCapacity(const size_t *uid, const size_t *vid, const double *cap, st
  * @param graph Graph structure in question
  * @return 0 of there was an error (edge not found, for example); otherwise, 1 if the flow value as successfully set.
  */
-int arraySetFlow(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
+short arraySetFlow(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
 
 /**
  * @brief Function to adjust the flow value of a given edge.
@@ -178,7 +178,7 @@ int arraySetFlow(const size_t *uid, const size_t *vid, const double *flow, struc
  * @param graph The graph structure in question
  * @return 0 if there was an error (such as the edge not found); otherwise, 1 if the flow value was successfully adjusted.
  */
-int arrayAddFlow(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
+short arrayAddFlow(const size_t *uid, const size_t *vid, const double *flow, struct graph_t *graph);
 
 
 /**
@@ -191,7 +191,7 @@ int arrayAddFlow(const size_t *uid, const size_t *vid, const double *flow, struc
  * @param callback Callback to be executed when graph has been reset.
  * @return 0 if there was an error during the reset; 1 if the reset completed;
  */
-int arrayResetGraph(struct graph_t *graph, void *args, void (*callback)(void));
+short arrayResetGraph(struct graph_t *graph, void *args, void (*callback)(void));
 
 /**
  * @brief Utility method to free up an array
@@ -201,7 +201,7 @@ int arrayResetGraph(struct graph_t *graph, void *args, void (*callback)(void));
  * @param arr 2D array to be cleared
  * @return 1 if successful; 0 if error.
  */
-int freeGraphArray(size_t arraylen, void** arrptr);
+short freeGraphArray(size_t arraylen, void** arrptr);
 
 
 #endif //GRAPHDATA_ARRAYOPS_H
